backend/requirements.txt
fastapi
uvicorn
opencv-python
pytesseract
python-multipart
python-dotenv
